+++ pkg/manifests/config_test.go
@@ -190,6 +190,30 @@ func TestEtcdDefaultsToDisabled(t *testing.T) {
 	}
 }
 
+func TestPromAdapterDedicatedSMsDefaultsToDisabled(t *testing.T) {
+	c, err := NewConfigFromString("")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if c.ClusterMonitoringConfiguration.K8sPrometheusAdapter.DedicatedServiceMonitors.Enabled {
+		t.Error("an empty configuration should have prometheus-adapter dedicated ServiceMonitors dislabled")
+	}
+	c, err = NewConfigFromString(`{"k8sPrometheusAdapter":{}}`)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if c.ClusterMonitoringConfiguration.K8sPrometheusAdapter.DedicatedServiceMonitors.Enabled {
+		t.Error("an empty k8sPrometheusAdapter configuration should have prometheus-adapter dedicated ServiceMonitors dislabled")
+	}
+	c, err = NewConfigFromString(`{"k8sPrometheusAdapter":{"dedicatedServiceMonitors":{}}}`)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if c.ClusterMonitoringConfiguration.K8sPrometheusAdapter.DedicatedServiceMonitors.Enabled {
+		t.Error("an empty dedicatedSericeMonitors configuration should have prometheus-adapter dedicated ServiceMonitors dislabled")
+	}
+}
+
 func TestHttpProxyConfig(t *testing.T) {
 	conf := `http:
   httpProxy: http://test.com